### 2021-11-19 #3927 @whedon
[PRE REVIEW]: pyscreener: A Python Wrapper for Computational Docking Software
**Submitting author:** @davidegraff (<a href="http://orcid.org/0000-0003-1250-3329">David E. Graff</a>)
**Repository:** <a href="https://github.com/coleygroup/pyscreener.git" target ="_blank">https://github.com/coleygroup/pyscreener.git</a>
**Version:** v1.1.1
