### 2021-11-24 #3950 @whedon
[REVIEW]: pyscreener: A Python Wrapper for Computational Docking Software
**Submitting author:** @davidegraff (<a href="http://orcid.org/0000-0003-1250-3329">David E. Graff</a>)
**Repository:** https://github.com/coleygroup/pyscreener.git
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.1.1

- [x] #4044
