### 2021-11-29 ada2abd
updated reading file bytes
