+++ tests/test_data.py
@@ -34,25 +34,12 @@ def test_score(smi):
 
     assert data.result.score == score
 
-# def test_file_bytes(smi):
-#     tempf = tempfile.TemporaryFile()
-#     tempf.write(b'Hello world!')
-
-#     data = CalculationData(smi, None, None, None, None, input_file = tempf)
-
-#     with open(tempf) as f:
-#         Bytes = f.read()
-#         print(Bytes)
-#         assert filecmp.cmp(data.input_file_bytes, Bytes)
-
 CONTENT = "testing for same memory"
 def test_create_file(tmp_path):
     d = tmp_path / "sub"
     d.mkdir()
     p = d / "hello.txt"
     p.write_text(CONTENT)
-    hello = p.read_text()
-    print(hello)
     assert p.read_text() == CONTENT
     
     data = CalculationData(smi, None, None, None, None, input_file = p)
