### 2021-12-06 5cabce9
extended testing for more file types
