### 2022-01-04 13fdb57
changed to relative path on moby dick and correct data type
