+++ tests/test_data.py
@@ -7,6 +7,8 @@ import pytest
 from pyscreener.docking import CalculationData, Result
 from pyscreener.exceptions import InvalidResultError, NotSimulatedError
 
+from pathlib import Path
+
 
 
 @pytest.fixture(
@@ -75,21 +77,13 @@ def test_input_large_file(tmp_path):
     data = CalculationData(smi, None, None, None, None, input_file = p)
     assert data.input_file_bytes == file_bytes
 
-
 def test_input_large_file():
-    with open('/home/mburlage/pyscreener/tests/mobydick.txt', 'rb') as f:
+    with open('tests/mobydick.txt', 'rb') as f:
         file_bytes = f.read()
         f.close()
-    data = CalculationData(smi, None, None, None, None, input_file = '/home/mburlage/pyscreener/tests/mobydick.txt')
+    data = CalculationData(smi, None, None, None, None, input_file = 'tests/mobydick.txt')
     assert data.input_file_bytes == file_bytes
-
-
-def test_input_large_file():
-    with open('/home/mburlage/pyscreener/tests/mobydick.txt', 'rb') as f:
-        file_bytes = f.read()
-        f.close()
-    data = CalculationData(smi, None, None, None, None, input_file = '/home/mburlage/pyscreener/tests/mobydick.txt')
-    assert data.input_file_bytes == file_bytes
+    
 
 CONTENT = "CCCCCCCC"
 def test_input_std_use_file(tmp_path):
