### 2022-01-11 #3950 @mikemhenry
@davidegraff I've got some more detailed comments about the installation and testing, but I found this package and paper to be excellent! I'll open up a few issues tomorrow (and let my computer run overnight to test out the scaling).
