### 2022-01-12 coleygroup/pyscreener#21 @rvhonorato
[JOSS review] Installation instructions
_Companion of openjournals/joss-reviews/issues/3950_

The install instructions are not very clear to me, but after following it step-by-step I tried to check it with:
```
$ pyscreener-check SCREEN_TYPE METADATA_TEMPLATE
```

Which seem like a nice feature, but its not clear what are the `SCREEN_TYPE` and `METADATA_TEMPLATE` and also there is no usage:

```
$ pyscreener-check -h
Traceback (most recent call last):
  File "/Users/rodrigo/software/anaconda3/envs/pyscreener_env/bin/pyscreener-check", line 8, in <module>
    sys.exit(check())
  File "/Users/rodrigo/repos/pyscreener/pyscreener/main.py", line 13, in check
    ps.check_env(sys.argv[1], json.loads(sys.argv[2]))
IndexError: list index out of range
```

Could you please clarify this step?

An additional note is that to install the packages you need to first add the `conda-forge` channel with `$ conda config --append channels conda-forge`
