### 2022-01-12 coleygroup/pyscreener#23 @rvhonorato
[JOSS review] Example usage
_Companion of openjournals/joss-reviews/issues/3950_

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Some exemple usage is provided in the paper and also in the README.md, but would be good to have it as a script in a new `examples/` folder. This is very helpful for users as a starting point to setup their own experiments and a quick way to experiment with the software capabilities.
