### 2022-01-12 coleygroup/pyscreener#24 @rvhonorato
[JOSS review] Community guidelines
_Companion of openjournals/joss-reviews/issues/3950_

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The repository already contains a `CODE_OF_CONDUCT.md` but would also be beneficial to have a `CONTRIBUTING.md` detailing how to contribute to the repository, also there should be issue templates for feature requests/bugs/questions.
