### 2022-01-12 coleygroup/pyscreener#25 @rvhonorato
[JOSS review] Functionality documentation - linting
_Companion of openjournals/joss-reviews/issues/3950_

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

The code is well documented but it does not adhere to pep8 style, my recommendation would be to enforce code style in the repository via github actions and solve the many issues:

`$ flake8 --ignore E501 --count > flake8.log`

[flake8.log](https://github.com/coleygroup/pyscreener/files/7855146/flake8.log)
