### 2022-01-12 coleygroup/pyscreener#26 @mikemhenry
[JOSS review] Installation Instruction Simplification 
I was able to follow the instructions and run the integration tests :tada:
The details on installing different docking backends was :ok_hand:
I do think the installation instructions could be simplified by installing everything from conda-forge:

`conda create -n NAME -c conda-forge python=3.8 pip openbabel openmm rdkit pdbfixer`

Then the extra step to install pdbfixer from source isn't necessary.

This is a secondary (but related) question:

Are you interested in getting this package on conda-forge? Then end users would only need to (in addtion to installing the backend of thier choice) do:

`conda install -c conda-forge pyscreener`

cc https://github.com/openjournals/joss-reviews/issues/3950
