### 2022-01-12 61a5f93
Fix codecov badge
The codecov badge was broken since it was pointing to `master` instead of `main`.
