### 2022-01-13 coleygroup/pyscreener#21 @mikemhenry
I too found that
`$ pyscreener-check SCREEN_TYPE METADATA_TEMPLATE` was a bit confusing with the location it appears in the documentation, since at that point I wasn't sure what a `SCREEN_TYPE` or  `METADATA_TEMPLATE` was, but that is explained later.

It may be better to direct users to test their setup with something like

```
$ pyscreener --config integration-tests/configs/test_vina.ini --smoke-test
Checking environment and metadata for "vina" screen
  Checking PATH and environment variables ... PASS
  Validating metadata ...  PASS
Environment is properly set up!
```

Since that seems to check everything and `--smoke-test` is a pretty good description.
