### 2022-01-13 coleygroup/pyscreener#26 @davidegraff
I think this is a good solution and I'll look to change this.

Re: the second question. I would definitely love to do that and think it could alleviate a decent amount of setup difficulty. I just didn't have time to dedicate to learning conda packaging. If you have any advice, I'd love to hear it!
