### 2022-01-13 coleygroup/pyscreener#26 @mikemhenry
Since this package is on pypi, there is some tooling that will auto-generate what is needed to package this on conda-forge. Since all the dependencies (well not the docking backends but that is okay) are on conda-forge, then it should be pretty easy :tm: Once we wrap up the JOSS review, I would be happy to help to get this on conda-forge
