### 2022-01-13 coleygroup/pyscreener#26 @davidegraff
conda package coming soon™️ ...

thanks so much for the offer to help!
