### 2022-02-09 #3950 @rvhonorato
Done on my part! I don't have the resources right now to test the performance and I've opened some issues about things that could be improved. 🎉
