### 2022-02-09 #3950 @richardjgowers
Yep that’s the usual process.

On Wed, Feb 9, 2022 at 15:53, Rodrigo V Honorato ***@***.***>
wrote:

> Fine with me! Probably all needs to be resolved before the review can be
> finished @richardjgowers <https://github.com/richardjgowers>?
>
