+++ .bumpversion.cfg
@@ -1,0 +1,8 @@
+[bumpversion]
+current_version = 1.1.1
+commit = True
+tag = True
+
+[bumpversion:file:setup.cfg]
+
+[bumpversion:file:pyscreener/__init__.py]
