+++ setup.cfg
@@ -1,12 +1,50 @@
-[bumpversion]
-current_version = 1.1.1
+[metadata]
+name = pyscreener
+version = 1.1.1
+author = david graff
+author_email = deg711@g.harvard.edu
+description = pythonic interface to virtual screening software
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = MIT
+classifiers = 
+    Framework :: Pytest
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: MIT License
+	Programming Language :: Python :: 3
+	Programming Language :: Python :: 3.8
+    Topic :: Scientific/Engineering :: Bio-Informatics
+    Topic :: Scientific/Engineering :: Chemistry
+url = https://github.com/coleygroup/pyscreener
+platforms = Linux, Mac OS-X, Unix
 
-[versioneer]
-VCS = git
-style = pep440
-versionfile_source = pyscreener/_version.py
-versionfile_build = pyscreener/_version.py
-tag_prefix = ''
+[options]
+include_package_data = True
+packages = find:
+python_requires = >=3.8
+install_requires = 
+	colorama
+	configargparse
+    h5py
+	numpy
+    pandas
+    ray[default]
+	scipy
+	tqdm
 
-[aliases]
+[options.entry_points]
+console_scripts = 
+	pyscreener = pyscreener.main:main
+    pyscreener-check = pyscreener.main:check
+
+[options.extras_require]
+dev =
+	black
+    bumpversion
+	flake8
+	pre-commit
+	pytest
+viz = 
+	matplotlib
+	seaborn
 test = pytest
