### 2022-02-25 4387b7b
use package metadata for versioning scheme
