+++ pyscreener/__init__.py
@@ -1,9 +1,13 @@
 """pyscreener
 pythonic interface to virtual screening software
 """
+from importlib.metadata import version, PackageNotFoundError
+
 from . import args
 from .docking import build_metadata, check_env, virtual_screen
 from .supply import LigandSupply
 
-from . import _version
-__version__ = _version.get_versions()['version']
+try:
+    __version__ = version("pyscreener")
+except PackageNotFoundError:
+    __version__ = "1.1.1"
