+++ setup.cfg
@@ -32,6 +32,12 @@ install_requires =
 	scipy
 	tqdm
 
+[options.packages.find]
+exclude =
+    pyscreener.dft
+    pyscreener.md
+    pyscreener.docking.dock.scripts
+    
 [options.entry_points]
 console_scripts = 
 	pyscreener = pyscreener.main:main
