+++ .bumpversion.cfg
@@ -6,3 +6,5 @@ tag = True
 [bumpversion:file:setup.cfg]
 
 [bumpversion:file:pyscreener/__init__.py]
+
+[bumpversion:file:pyscreener/args.py]
