+++ pyscreener/postprocessing/cluster.py
@@ -16,14 +16,16 @@ from sklearn.cluster import MiniBatchKMeans
 
 from pyscreener.postprocessing import fingerprints
 
-def cluster(d_smi_score: Dict[str, Optional[float]],
-            name: str = 'clusters', path: str = '.', **kwargs) -> List[Dict]:
+
+def cluster(
+    d_smi_score: Dict[str, Optional[float]], name: str = "clusters", path: str = ".", **kwargs
+) -> List[Dict]:
     d_smi_cid = cluster_smis(d_smi_score.keys(), len(d_smi_score), **kwargs)
 
-    clusters_csv = (Path(path)/f'{name}_clusters').with_suffix('.csv')
-    with open(clusters_csv, 'w') as fid:
+    clusters_csv = (Path(path) / f"{name}_clusters").with_suffix(".csv")
+    with open(clusters_csv, "w") as fid:
         writer = csv.writer(fid)
-        writer.writerow(['smiles', 'cluster_id'])
+        writer.writerow(["smiles", "cluster_id"])
         writer.writerows(d_smi_cid.items())
 
     d_cluster_smi_score = defaultdict(dict)
@@ -33,10 +35,16 @@ def cluster(d_smi_score: Dict[str, Optional[float]],
 
     return list(d_cluster_smi_score.values())
 
-def cluster_smis(smis: Iterable[str], n_mols: int, *,
-                 n_cluster: int = 10, 
-                 path: str = '.', name: str = 'fps', 
-                 **kwargs) -> Dict[str, int]:
+
+def cluster_smis(
+    smis: Iterable[str],
+    n_mols: int,
+    *,
+    n_cluster: int = 10,
+    path: str = ".",
+    name: str = "fps",
+    **kwargs,
+) -> Dict[str, int]:
     """Cluster the SMILES strings
 
     Parameters
@@ -56,18 +64,18 @@ def cluster_smis(smis: Iterable[str], n_mols: int, *,
     -------
     d_smi_cid : Dict[str, int]
         a mapping from SMILES string to cluster ID
-    
+
     See also
     --------
     fingerprints.gen_fps_h5
     """
-    fps_h5, invalid_idxs = fingerprints.gen_fps_h5(
-        smis, n_mols, path=path, name=name, **kwargs)
+    fps_h5, invalid_idxs = fingerprints.gen_fps_h5(smis, n_mols, path=path, name=name, **kwargs)
 
     smis = [smi for i, smi in enumerate(smis) if i not in invalid_idxs]
     cids = cluster_fps_h5(fps_h5, n_cluster)
-    
+
     return dict(zip(smis, cids))
+
 
 def cluster_fps_h5(fps_h5: str, n_cluster: int = 10) -> List[int]:
     """Cluster the feature matrix of fingerprints in fps_h5
@@ -89,24 +97,23 @@ def cluster_fps_h5(fps_h5: str, n_cluster: int = 10) -> List[int]:
     """
     begin = timeit.default_timer()
 
-    with h5py.File(fps_h5, 'r') as h5f:
-        fps = h5f['fps']
+    with h5py.File(fps_h5, "r") as h5f:
+        fps = h5f["fps"]
         chunk_size = fps.chunks[0]
 
         ITER = 1000
         BATCH_SIZE = min(1000, len(fps))
         clusterer = MiniBatchKMeans(n_clusters=n_cluster, batch_size=BATCH_SIZE)
-        
+
         for _ in range(ITER):
             rand_idxs = sorted(sample(range(len(fps)), BATCH_SIZE))
             batch_fps = fps[rand_idxs]
             clusterer.partial_fit(batch_fps)
 
[…]
