+++ pyscreener/preprocessing/filter.py
@@ -8,26 +8,31 @@ from rdkit import Chem
 from rdkit.Chem import QED
 from tqdm import tqdm
 
+
 def filter_ligands(ligands: str, **kwargs) -> Tuple[List[str], Optional[List[str]]]:
     if isinstance(ligands, str):
         p_ligand = Path(ligands)
 
-        if p_ligand.suffix == '.csv':
+        if p_ligand.suffix == ".csv":
             return filter_csv(ligands, **kwargs)
-        if p_ligand.suffix in {'.sdf', '.smi'}:
+        if p_ligand.suffix in {".sdf", ".smi"}:
             return filter_supply(ligands, **kwargs)
-        
+
         return [ligands], None
 
     elif isinstance(ligands, Sequence):
         return filter_smis(ligands, **kwargs)
-    
+
     raise TypeError('argument "ligand" must be of type str or Sequence[str]!')
 
+
 def filter_mols(
-    mols: List[Chem.Mol], names: Optional[List[str]] = None,
-    max_atoms: int = 1000, max_weight: float = 10000.,
-    max_logP: float = 10., **kwargs
+    mols: List[Chem.Mol],
+    names: Optional[List[str]] = None,
+    max_atoms: int = 1000,
+    max_weight: float = 10000.0,
+    max_logP: float = 10.0,
+    **kwargs,
 ) -> Tuple[List[str], Optional[List[str]]]:
     """Filter a list of molecules according to input critera
 
@@ -50,13 +55,12 @@ def filter_mols(
         were originally supplied
     """
     names = names or []
-    
+
     smis_filtered = []
     names_filtered = []
 
     if names:
-        for mol, name in tqdm(zip(mols, names), total=len(mols),
-                            desc='Filtering mols', unit='mol'):
+        for mol, name in tqdm(zip(mols, names), total=len(mols), desc="Filtering mols", unit="mol"):
             if mol.GetNumHeavyAtoms() > max_atoms:
                 continue
 
@@ -70,8 +74,7 @@ def filter_mols(
             names_filtered.append(name)
     else:
         names_filtered = None
-        for mol in tqdm(mols, total=len(mols),
-                        desc='Filtering mols', unit='mol'):
+        for mol in tqdm(mols, total=len(mols), desc="Filtering mols", unit="mol"):
             if mol.GetNumHeavyAtoms() > max_atoms:
                 continue
 
@@ -85,44 +88,50 @@ def filter_mols(
 
     return smis_filtered, names_filtered
 
-def filter_smis(smis: List[str], names: Optional[List[str]] = None,
-                **kwargs) -> Tuple[List[str], Optional[List[str]]]:
-    mols = [mol_from_smi(smi)
-            for smi in tqdm(smis, desc='Reading in mols', unit='mol')]
+
+def filter_smis(
+    smis: List[str], names: Optional[List[str]] = None, **kwargs
+) -> Tuple[List[str], Optional[List[str]]]:
+    mols = [mol_from_smi(smi) for smi in tqdm(smis, desc="Reading in mols", unit="mol")]
 
     return filter_mols(mols, names, **kwargs)
 
-def filter_csv(csvfile: str, title_line: bool = True,
-               smiles_col: int = 0, name_col: Optional[int] = None,
-               **kwargs) -> Tuple[List[str], Optional[List[str]]]:
+
+def filter_csv(
+    csvfile: str,
+    title_line: bool = True,
+    smiles_col: int = 0,
+    name_col: Optional[int] = None,
+    **kwargs,
+) -> Tuple[List[str], Optional[List[str]]]:
     with open(csvfile) as fid:
         reader = csv.reader(fid)
         if title_line:
             next(reader)
[…]
