+++ pyscreener/docking/result.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from typing import Optional
 
+
 @dataclass
 class Result:
     smiles: str
