### 2022-02-25 coleygroup/pyscreener#25 @davidegraff
this should be fixed in the recent PR- closing for now!
