### 2022-02-25 c7494f9
update `environment.yml` and `setup.cfg` to add `pdbfixer` from conda
