+++ setup.cfg
@@ -28,6 +28,7 @@ install_requires =
     h5py
 	numpy
     pandas
+    pdbfixer
     ray[default]
 	scipy
 	tqdm
