+++ requirements.txt
@@ -1,11 +1,0 @@
-colorama
-configargparse
-h5py
-numpy
-ray
-pandas
-pytest
-scikit_learn
-scipy
-seaborn
-tqdm
