### 2022-02-25 fb4b0e0
redirect users to use `--smoke-test` flag
