### 2022-02-25 coleygroup/pyscreener#21 @davidegraff
hi @rvhonorato and @rvhonorato,

this issue has been addressed in the recent commits to `main`: c7494f9fb44d0e13d6a8f53a32aa8550cad46ae3..fb4b0e090468937e6bf7f8274b0e173b43fd0de1

I'm closing this for now, but feel free to reopen if it's not sufficient!
