+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,12 +1,20 @@
 ## Description
-Provide a brief description of the PR's purpose here.
+Include a brief summary of the bug/feature/etc. that this PR seeks to address
 
-## Todos
-Notable points that this PR has either accomplished or will accomplish.
-  - [ ] TODO 1
+## Example / Current workflow
+Include a sample workflow to either **(a)** reproduce the bug with current codebase or **(b)** showcase the deficiency does this PR seeks to address
+
+## Bugfix / Desired workflow
+Include either **(a)** the same workflow from above with the correct output produced via this PR **(b)** some (pseudo)code containing the new workflow that this PR will (seek to) implement
 
 ## Questions
-- [ ] Question1
+If there are open questions about implementation strategy or scope of the PR, include them here
+
+## Relevant issues
+If appropriate, please tag them here and include a quick summary
+
+## checklist
+- [ ] (if appropriate) unit tests added: `pytest`
 
 ## Status
 - [ ] Ready to go
