+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,19 @@
+---
+name: Question
+about: Have a question about how to use pyscreener?
+title: "[SETUP]: "
+labels: question
+assignees: ''
+
+---
+
+**Question**
+What are you trying to do? Include details as to your intended goal, what sort of screen you're trying to accomplish, and what problems you've been having
+
+**Environment**
+- python version: `python -V`
+- package versions: `conda list`
+- OS
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
