### 2022-02-25 coleygroup/pyscreener#24 @davidegraff
addressed in #32
