### 2022-02-25 coleygroup/pyscreener#23 @davidegraff
addressed in #33 !
