+++ pyproject.toml
@@ -5,6 +5,17 @@ build-backend = "setuptools.build_meta"
 [tool.setuptools_scm]
 write_to = "pyscreener/_version.py"
 
+[tool.pytest.ini_options]
+addopts = "--cov pyscreener"
+
+[tool.coverage.run]
+branch = true
+omit = [
+    "pyscreener/dft/*",
+    "pyscreener/md/*",
+    "pyscreener/docking/dock/scripts/*"
+]
+
 [tool.black]
 line-length = 100
 target-version = ["py38"]
