### 2022-03-04 #3950 @richardjgowers
@davidegraff thanks for your patience, I think we're ready to move forwards.  Can you archive a release of the software on zenodo (or similar) and report the DOI back here?  Also is the version still v1.1.1?
