+++ .github/workflows/draft-pdf.yaml
@@ -1,9 +1,11 @@
+name: paper
+
 on: 
   push:
     branches:
       - "joss-paper"
-    paths:
-      - "paper.md"
+    # paths:
+    #   - "paper.md"
 
 jobs:
   paper:
@@ -12,7 +14,7 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-        
+
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
