### 2022-03-04 #3950 @davidegraff
@richardjgowers the version is still 1.1.1 and the DOI from zenodo is 10.5281/zenodo.6327783
