### 2022-03-14 effbc88
rename `CalculationData` -> `Simulation` and update packaging
