+++ pyscreener/warnings.py
@@ -1,0 +1,8 @@
+class ConformerWarning(Warning):
+    pass
+
+class ChargeWarning(Warning):
+    pass
+
+class SimulationFailureWarning(Warning):
+    pass
