+++ pyscreener/docking/vina/runner.py
@@ -5,6 +5,7 @@ import shutil
 import subprocess as sp
 import sys
 from typing import List, Optional, Tuple, Union
+import warnings
 
 from openbabel import pybel
 from rdkit.Chem import AllChem as Chem
@@ -12,7 +13,8 @@ import ray
 
 from pyscreener import utils
 from pyscreener.exceptions import MissingExecutableError
-from pyscreener.docking.data import CalculationData
+from pyscreener.warnings import ChargeWarning, ConformerWarning, SimulationFailureWarning
+from pyscreener.docking.sim import Simulation
 from pyscreener.docking.runner import DockingRunner
 from pyscreener.docking.result import Result
 from pyscreener.docking.vina.metadata import VinaMetadata
@@ -27,14 +29,14 @@ if shutil.which("prepare_receptor") is None:
 
 class VinaRunner(DockingRunner):
     @staticmethod
-    def prepare(data: CalculationData) -> CalculationData:
+    def prepare(data: Simulation) -> Simulation:
         data = VinaRunner.prepare_receptor(data)
         data = VinaRunner.prepare_ligand(data)
 
         return data
 
     @staticmethod
-    def prepare_receptor(data: CalculationData) -> CalculationData:
+    def prepare_receptor(data: Simulation) -> Simulation:
         """Prepare a receptor PDBQT file from its input file
 
         Parameters
@@ -63,24 +65,25 @@ class VinaRunner(DockingRunner):
         return data
 
     @staticmethod
-    def prepare_and_run(data: CalculationData) -> CalculationData:
+    def prepare_and_run(data: Simulation) -> Simulation:
         VinaRunner.prepare_ligand(data)
         VinaRunner.run(data)
 
         return data
 
     @staticmethod
-    def prepare_ligand(data: CalculationData) -> CalculationData:
-        if data.smi is not None:
-            VinaRunner.prepare_from_smi(data)
-        else:
-            VinaRunner.prepare_from_file(data)
-
-        return data
+    def prepare_ligand(data: Simulation) -> bool:
+        return (
+            VinaRunner.prepare_from_smi(data) if data.smi is not None
+            else VinaRunner.prepare_from_file(data)
+        )
 
     @staticmethod
-    def prepare_from_smi(data: CalculationData) -> CalculationData:
-        """Prepare an input ligand file from the ligand's SMILES string
+    def prepare_from_smi(data: Simulation) -> bool:
+        """Prepare the ligand PDQBT file from its SMILES string
+
+        Sets the `prepared_ligand` attribute of the metadata. If ligand preparation fails for any
+        reason, set the value to `None` and return.
 
         Parameters
         ----------
@@ -88,31 +91,47 @@ class VinaRunner(DockingRunner):
 
         Returns
         -------
-        CalculationData
+        bool
+            whether the ligand preparation succeeded
         """
         pdbqt = Path(data.in_path) / f"{data.name}.pdbqt"
 
-        mol = Chem.AddHs(Chem.MolFromSmiles(data.smi))
-        Chem.EmbedMolecule(mol)
-        Chem.MMFFOptimizeMolecule(mol)
+        mol = Chem.MolFromSmiles(data.smi)
+        if mol is None:
+            data.metadata.prepared_ligand = None
+            return False
+
+        mol = Chem.AddHs(mol)
+
+        try:
+            Chem.EmbedMolecule(mol)
+            Chem.MMFFOptimizeMolecule(mol)
+        except ValueError:
+            warnings.warn("Could not generate 3D conformer of molecule!", ConformerWarning)
 
         try:
             mol = pybel.readstring("mol", Chem.MolToMolBlock(mol))
-            # mol.make3D()
-            # mol.addh()
+        except IOError:
[…]
