+++ pyscreener/docking/sim.py
@@ -4,12 +4,12 @@ from typing import Optional, Tuple, Union
 
 from pyscreener.exceptions import InvalidResultError, NotSimulatedError
 from pyscreener.utils import ScoreMode
-from pyscreener.docking.metadata import CalculationMetadata
+from pyscreener.docking.metadata import SimulationMetadata
 from pyscreener.docking.result import Result
 
 
 @dataclass(repr=True, eq=False)
-class CalculationData:
+class Simulation:
     """
 
     Attributes
@@ -59,7 +59,7 @@ class CalculationData:
     receptor: str
     center: Tuple[float, float, float]
     size: Tuple[float, float, float]
-    metadata: CalculationMetadata
+    metadata: SimulationMetadata
     ncpu: int = 1
     name: str = "ligand"
     input_file: Optional[str] = None
