+++ MANIFEST.in
@@ -1,6 +1,0 @@
-include LICENSE
-include MANIFEST.in
-include CODE_OF_CONDUCT.md
-
-graft pyscreener
-global-exclude *.py[cod] __pycache__ *.so

[…]
