+++ pyscreener/docking/vina/runner.py
@@ -29,14 +29,14 @@ if shutil.which("prepare_receptor") is None:
 
 class VinaRunner(DockingRunner):
     @staticmethod
-    def prepare(data: Simulation) -> Simulation:
-        data = VinaRunner.prepare_receptor(data)
-        data = VinaRunner.prepare_ligand(data)
+    def prepare(sim: Simulation) -> Simulation:
+        sim = VinaRunner.prepare_receptor(sim)
+        sim = VinaRunner.prepare_ligand(sim)
 
-        return data
+        return sim
 
     @staticmethod
-    def prepare_receptor(data: Simulation) -> Simulation:
+    def prepare_receptor(sim: Simulation) -> Simulation:
         """Prepare a receptor PDBQT file from its input file
 
         Parameters
@@ -49,37 +49,37 @@ class VinaRunner(DockingRunner):
         receptor_pdbqt : Optional[str]
             the filepath of the resulting PDBQT file. None if preparation failed
         """
-        name = Path(data.receptor).with_suffix(".pdbqt").name
-        receptor_pdbqt = Path(data.in_path) / name
+        name = Path(sim.receptor).with_suffix(".pdbqt").name
+        receptor_pdbqt = Path(sim.in_path) / name
 
-        argv = ["prepare_receptor", "-r", data.receptor, "-o", receptor_pdbqt]
+        argv = ["prepare_receptor", "-r", sim.receptor, "-o", receptor_pdbqt]
         try:
             ret = sp.run(argv, stderr=sp.PIPE)
             ret.check_returncode()
         except sp.SubprocessError:
-            print(f'ERROR: failed to convert "{data.receptor}"', file=sys.stderr)
+            print(f'ERROR: failed to convert "{sim.receptor}"', file=sys.stderr)
             print(ret.stderr.decode("utf-8"))
             return None
 
-        data.metadata.prepared_receptor = receptor_pdbqt
-        return data
+        sim.metadata.prepared_receptor = receptor_pdbqt
+        return sim
 
     @staticmethod
-    def prepare_and_run(data: Simulation) -> Simulation:
-        VinaRunner.prepare_ligand(data)
-        VinaRunner.run(data)
+    def prepare_and_run(sim: Simulation) -> Simulation:
+        VinaRunner.prepare_ligand(sim)
+        VinaRunner.run(sim)
 
-        return data
+        return sim
 
     @staticmethod
-    def prepare_ligand(data: Simulation) -> bool:
+    def prepare_ligand(sim: Simulation) -> bool:
         return (
-            VinaRunner.prepare_from_smi(data) if data.smi is not None
-            else VinaRunner.prepare_from_file(data)
+            VinaRunner.prepare_from_smi(sim) if sim.smi is not None
+            else VinaRunner.prepare_from_file(sim)
         )
 
     @staticmethod
-    def prepare_from_smi(data: Simulation) -> bool:
+    def prepare_from_smi(sim: Simulation) -> bool:
         """Prepare the ligand PDQBT file from its SMILES string
 
         Sets the `prepared_ligand` attribute of the metadata. If ligand preparation fails for any
@@ -87,18 +87,18 @@ class VinaRunner(DockingRunner):
 
         Parameters
         ----------
-        data: CalculationData
+        sim : CalculationData
 
         Returns
         -------
         bool
             whether the ligand preparation succeeded
         """
-        pdbqt = Path(data.in_path) / f"{data.name}.pdbqt"
+        pdbqt = Path(sim.in_path) / f"{sim.name}.pdbqt"
 
-        mol = Chem.MolFromSmiles(data.smi)
+        mol = Chem.MolFromSmiles(sim.smi)
         if mol is None:
-            data.metadata.prepared_ligand = None
+            sim.metadata.prepared_ligand = None
             return False
 
[…]
