+++ pyscreener/docking/dock/runner.py
@@ -33,21 +33,21 @@ for f in (VDW_DEFN_FILE, FLEX_DEFN_FILE, FLEX_DRIVE_FILE, DOCK):
     if not f.exists():
         raise MisconfiguredDirectoryError(
             f'$DOCK6 directory not configured properly! DOCK6 path is set as "{DOCK6}", but there '
-            f'is no "{f.name}" located under the "{f.parents}" directory. '
+            f'is no "{f.name}" located under the "{f.parents[0].name}" directory. '
             "See https://github.com/coleygroup/pyscreener#specifying-an-environment-variable for more information."
         )
 
 
 class DOCKRunner(DockingRunner):
     @staticmethod
-    def prepare(data: Simulation) -> Simulation:
-        data = DOCKRunner.prepare_receptor(data)
-        data = DOCKRunner.prepare_ligand(data)
+    def prepare(sim: Simulation) -> Simulation:
+        sim = DOCKRunner.prepare_receptor(sim)
+        sim = DOCKRunner.prepare_ligand(sim)
 
-        return data
+        return sim
 
     @staticmethod
-    def prepare_receptor(data: Simulation) -> Simulation:
+    def prepare_receptor(sim: Simulation) -> Simulation:
         """Prepare the files necessary to dock ligands against the input receptor using this
         Screener's parameters
 
@@ -66,111 +66,108 @@ class DOCKRunner(DockingRunner):
         None
             if receptor preparation fails at any point
         """
-        rec_mol2 = utils.prepare_mol2(data.receptor, data.in_path)
-        rec_pdb = utils.prepare_pdb(data.receptor, data.in_path)
+        rec_mol2 = utils.prepare_mol2(sim.receptor, sim.in_path)
+        rec_pdb = utils.prepare_pdb(sim.receptor, sim.in_path)
 
         if rec_mol2 is None or rec_pdb is None:
-            return data
+            return sim
 
-        rec_dms = utils.prepare_dms(rec_pdb, data.metadata.probe_radius, data.in_path)
+        rec_dms = utils.prepare_dms(rec_pdb, sim.metadata.probe_radius, sim.in_path)
         if rec_dms is None:
-            return data
+            return sim
 
         rec_sph = utils.prepare_sph(
             rec_dms,
-            data.metadata.steric_clash_dist,
-            data.metadata.min_radius,
-            data.metadata.max_radius,
-            data.in_path,
+            sim.metadata.steric_clash_dist,
+            sim.metadata.min_radius,
+            sim.metadata.max_radius,
+            sim.in_path,
         )
         if rec_sph is None:
-            return data
+            return sim
 
         rec_sph = utils.select_spheres(
             rec_sph,
-            data.metadata.sphere_mode,
-            data.center,
-            data.size,
-            data.metadata.docked_ligand_file,
-            data.metadata.buffer,
-            data.in_path,
+            sim.metadata.sphere_mode,
+            sim.center,
+            sim.size,
+            sim.metadata.docked_ligand_file,
+            sim.metadata.buffer,
+            sim.in_path,
         )
 
         rec_box = utils.prepare_box(
             rec_sph,
-            data.center,
-            data.size,
-            data.metadata.enclose_spheres,
-            data.metadata.buffer,
-            data.in_path,
+            sim.center,
+            sim.size,
+            sim.metadata.enclose_spheres,
+            sim.metadata.buffer,
+            sim.in_path,
         )
         if rec_box is None:
-            return data
+            return sim
 
-        grid_stem = utils.prepare_grid(rec_mol2, rec_box, data.in_path)
[…]
