### 2022-03-14 dd82e07
pass back `Result` instead of full `Simulation`
