+++ pyscreener/warnings.py
@@ -1,8 +1,10 @@
 class ConformerWarning(Warning):
     pass
 
+
 class ChargeWarning(Warning):
     pass
+
 
 class SimulationFailureWarning(Warning):
     pass
