+++ readthedocs.yml
@@ -1,15 +1,0 @@
-# readthedocs.yml
-
-version: 2
-
-build:
-  image: latest
-
-python:
-  version: 3.6
-  install:
-    - method: pip
-      path: .
-
-conda:
-  environment: docs/requirements.yaml
