+++ docs/requirements.yaml
@@ -1,26 +1,0 @@
-name: docs
-
-channels:
-  - conda-forge
-  - defaults
-
-dependencies:
-  - python=3.6
-  - pip
-
-  - hdf5
-  - importlib_resources
-  - matplotlib
-  - mpi4py
-  - openbabel
-  - pandas
-  - omnia::pdbfixer
-  - pytest
-  - python=3.6
-  - rdkit
-  - retrying
-  - scikit-learn
-  - scipy
-  - seaborn
-  - tqdm
-
