+++ devtools/conda-envs/requirements.txt
@@ -1,14 +1,0 @@
-codecov
-colorama
-configargparse
-git+https://github.com/openmm/pdbfixer.git
-h5py
-numpy
-ray
-pandas
-pytest
-pytest-cov
-scikit-learn
-scipy
-seaborn
-tqdm
