+++ MANIFEST.IN
@@ -1,5 +1,0 @@
-global-exclude .*
-recursive-exclude .github *
-prune devtools
-prune examples
-prune integration-tests
