+++ pyscreener/docking/runner.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Optional, Sequence
-from pyscreener.docking.result import Result
 
+from pyscreener.docking.result import Result
 from pyscreener.docking.sim import Simulation
 from pyscreener.docking.metadata import SimulationMetadata
 
@@ -25,8 +25,8 @@ class DockingRunner(ABC):
     @staticmethod
     @abstractmethod
     def prepare_and_run(data: Simulation) -> Optional[Result]:
-        """Prepare the receptor and ligand files then run the given simulation. Roughly equivlaent
-        to `prepare_*()` followed by `run()` but doesn't return the the scores of the conformers"""
+        """Prepare the ligand file then run the given simulation. Roughly equivlaent to `prepare_ligand()` followed by `run()` but returns the Result object for the Simulation
+        rather than the scores of the conformers"""
 
     @staticmethod
     def validate_metadata(metadata: SimulationMetadata):
