+++ pyscreener/utils/utils.py
@@ -90,7 +90,6 @@ def calc_score(
 
 def reduce_scores(
     S: np.ndarray,
-    repeat_score_mode: ScoreMode = ScoreMode.BEST,
     ensemble_score_mode: ScoreMode = ScoreMode.BEST,
     k: int = 1,
 ) -> Optional[float]:
@@ -99,12 +98,10 @@ def reduce_scores(
     Parameters
     ----------
     S : np.ndarray
-        an `n x r x t` array of docking scores, where n is the number of ligands that were docked,
+        an `n x r` array of docking scores, where n is the number of ligands that were docked,
         r is the number of receptors each ligand was docked against, and t is the number of repeated
         docking attempts against each receptor, and each value is the docking score calculated for
         the given run
-    repeat_score_mode : ScoreMode, default=ScoreMode.BEST,
-        the mode used to calculate the overall score for from repeated runs
     ensemble_score_mode : ScoreMode, default=ScoreMode.BEST,
         the mode used to calculate the overall score for a given ensemble of receptors
     k : int, default=1
@@ -117,17 +114,6 @@ def reduce_scores(
     """
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore", r"All-NaN (slice|axis) encountered")
-
-        if repeat_score_mode == ScoreMode.BEST:
-            S = np.nanmin(S, axis=2)
-        elif repeat_score_mode == ScoreMode.AVG:
-            S = np.nanmean(S, axis=2)
-        elif repeat_score_mode == ScoreMode.BOLTZMANN:
-            S_e = np.exp(-S)
-            Z = S_e / np.nansum(S_e, axis=2)[:, :, None]
-            S = np.nansum((S * Z), axis=2)
-        elif repeat_score_mode == ScoreMode.TOP_K:
-            S = np.nanmean(np.sort(S, axis=2)[:, :k], axis=2)
 
         if ensemble_score_mode == ScoreMode.BEST:
             S = np.nanmin(S, axis=1)
