+++ pyscreener/docking/dock/exceptions.py
@@ -1,0 +1,17 @@
+from pyscreener.exceptions import ReceptorPreparationError
+
+
+class SurfaceGenerationError(ReceptorPreparationError):
+    pass
+
+
+class SphereGenerationError(ReceptorPreparationError):
+    pass
+
+
+class BoxGenerationError(ReceptorPreparationError):
+    pass
+
+
+class GridGenerationError(ReceptorPreparationError):
+    pass
