+++ pyscreener/docking/dock/utils.py
@@ -7,8 +7,18 @@ import shutil
 import subprocess as sp
 import sys
 from typing import Optional, Tuple
+from pyscreener.docking.dock.exceptions import (
+    BoxGenerationError,
+    GridGenerationError,
+    SphereGenerationError,
+    SurfaceGenerationError,
+)
 
-from pyscreener.exceptions import MissingEnvironmentVariableError, MisconfiguredDirectoryError
+from pyscreener.exceptions import (
+    MissingEnvironmentVariableError,
+    MisconfiguredDirectoryError,
+    ReceptorPreparationError,
+)
 from pyscreener.utils import AutoName
 
 with resources.path("pyscreener.docking.dock", ".") as p_module:
@@ -45,74 +55,74 @@ class SphereMode(AutoName):
     LIGAND = auto()
 
 
-def prepare_receptor(
-    receptor: str,
-    probe_radius: float = 1.4,
-    steric_clash_dist: float = 0.0,
-    min_radius: float = 1.4,
-    max_radius: float = 4.0,
-    sphere_mode: SphereMode = SphereMode.LARGEST,
-    center: Optional[Tuple[float, float, float]] = None,
-    size: Tuple[float, float, float] = (10.0, 10.0, 10.0),
-    docked_ligand_file: Optional[str] = None,
-    buffer: float = 10.0,
-    enclose_spheres: bool = True,
-    path: str = ".",
-) -> Optional[Tuple[str, str]]:
-    """Prepare the DOCK input files corresponding to the given receptor
+# def prepare_receptor(
+#     receptor: str,
+#     probe_radius: float = 1.4,
+#     steric_clash_dist: float = 0.0,
+#     min_radius: float = 1.4,
+#     max_radius: float = 4.0,
+#     sphere_mode: SphereMode = SphereMode.LARGEST,
+#     center: Optional[Tuple[float, float, float]] = None,
+#     size: Tuple[float, float, float] = (10.0, 10.0, 10.0),
+#     docked_ligand_file: Optional[str] = None,
+#     buffer: float = 10.0,
+#     enclose_spheres: bool = True,
+#     path: str = ".",
+# ) -> Optional[Tuple[str, str]]:
+#     """Prepare the DOCK input files corresponding to the given receptor
 
-    Parameters
-    ----------
-    receptor : str
-        the filepath of a file containing a receptor. Must be in a file that
-        is readable by Chimera
-    center : Tuple[float, float, float]
-        the x-, y-, and z-coordinates of the center of the docking box
-    size : Tuple[float, float, float] (Default = (20, 20, 20))
-        the x-, y-, and z-radii of the docking box
-    docked_ligand_file : Optional[str] (Default = None)
-        the filepath of a file containing the coordinates of a docked ligand
-    use_largest : bool (Default = False)
-        whether to use the largest cluster of spheres when selecting spheres
-    buffer : float (Default = 10.)
-        the amount of buffer space to be added around the docked ligand when
-        selecting spheres and when constructing the docking box if
-        enclose_spheres is True
-    enclose_spheres : bool (Default = True)
-        whether to calculate the docking box by enclosing the selected spheres
-        or to use an input center and radii
+#     Parameters
+#     ----------
+#     receptor : str
+#         the filepath of a file containing a receptor. Must be in a file that
+#         is readable by Chimera
+#     center : Tuple[float, float, float]
+#         the x-, y-, and z-coordinates of the center of the docking box
+#     size : Tuple[float, float, float] (Default = (20, 20, 20))
+#         the x-, y-, and z-radii of the docking box
+#     docked_ligand_file : Optional[str] (Default = None)
[…]
