### 2022-03-14 2a00954
catch exception for now
