+++ pyscreener/__init__.py
@@ -1,13 +1,14 @@
 """pyscreener
 pythonic interface to virtual screening software
 """
+from modulefinder import Module
 from . import args
 from .docking import build_metadata, check_env, virtual_screen
 from .supply import LigandSupply
 
 try:
-    from . import _version
+    from ._version import version
 
-    __version__ = _version.version
+    __version__ = version
 except ModuleNotFoundError:
     __version__ = "1.1.1"
