### 2022-03-15 d3f2aba
add multithreaded classmethod
