### 2022-03-15 e403bc0
switch to using param dictionaries for grid and docking
