+++ pyscreener/docking/dock/runner.py
@@ -44,6 +44,10 @@ for f in (VDW_DEFN_FILE, FLEX_DEFN_FILE, FLEX_DRIVE_FILE, DOCK):
 
 
 class DOCKRunner(DockingRunner):
+    @classmethod
+    def is_multithreaded(cls) -> bool:
+        return False
+
     @staticmethod
     def prepare(sim: Simulation) -> Simulation:
         _ = DOCKRunner.prepare_receptor(sim)
@@ -65,16 +69,19 @@ class DOCKRunner(DockingRunner):
         -------
         Simulation
         """
+        in_path = sim.in_path / "receptors"
+        in_path.mkdir(parents=True, exist_ok=True)
+
         try:
-            rec_mol2 = utils.prepare_mol2(sim.receptor, sim.in_path)
-            rec_pdb = utils.prepare_pdb(sim.receptor, sim.in_path)
-            rec_dms = utils.prepare_dms(rec_pdb, sim.metadata.probe_radius, sim.in_path)
+            rec_mol2 = utils.prepare_mol2(sim.receptor, in_path)
+            rec_pdb = utils.prepare_pdb(sim.receptor, in_path)
+            rec_dms = utils.prepare_dms(rec_pdb, sim.metadata.probe_radius, in_path)
             rec_sph = utils.prepare_sph(
                 rec_dms,
                 sim.metadata.steric_clash_dist,
                 sim.metadata.min_radius,
                 sim.metadata.max_radius,
-                sim.in_path,
+                in_path,
             )
             rec_sph = utils.select_spheres(
                 rec_sph,
@@ -83,7 +90,7 @@ class DOCKRunner(DockingRunner):
                 sim.size,
                 sim.metadata.docked_ligand_file,
                 sim.metadata.buffer,
-                sim.in_path,
+                in_path,
             )
             rec_box = utils.prepare_box(
                 rec_sph,
@@ -91,11 +98,12 @@ class DOCKRunner(DockingRunner):
                 sim.size,
                 sim.metadata.enclose_spheres,
                 sim.metadata.buffer,
-                sim.in_path,
+                in_path,
             )
-            grid_stem = utils.prepare_grid(rec_mol2, rec_box, sim.in_path)
+            grid_stem = utils.prepare_grid(rec_mol2, rec_box, in_path)
         except ReceptorPreparationError:
-            return None  # should think about whether to handle or just raise
+            raise
+            # return None  # should think about whether to handle or just raise
 
         sim.metadata.prepared_receptor = rec_sph, grid_stem
         return sim
@@ -134,6 +142,8 @@ class DOCKRunner(DockingRunner):
         if mol is None:
             return False
 
+        mol = Chem.AddHs(mol)
+
         try:
             Chem.EmbedMolecule(mol)
             Chem.MMFFOptimizeMolecule(mol)
@@ -161,17 +171,17 @@ class DOCKRunner(DockingRunner):
         fmt = Path(sim.input_file).suffix.strip(".")
 
         mol = next(pybel.readfile(fmt, sim.input_file))
+        p_mol2 = Path(sim.in_path) / f"{mol.title or sim.name}.mol2"
 
-        mol2 = Path(sim.in_path) / f"{mol.title or sim.name}.mol2"
         sim.smi = mol.write()
+        mol.addh()
         try:
-            mol.addh()
             mol.calccharges(model="gasteiger")
         except Exception:
-            pass
+            warnings.warn("Could not calculate charges for ligand!", ChargeWarning)
 
-        mol.write(format="mol2", filename=str(mol2), overwrite=True, opt={"h": None})
-        sim.metadata.prepared_ligand = mol2
[…]
