### 2022-03-15 #3950 @kthyng
Hi @davidegraff! I am taking over publishing from here. Can you update the metadata in your zenodo archive so the title and author list match your JOSS paper?
