### 2022-03-16 #3950 @davidegraff
Hi @kthyng ! Thanks for taking over- I believe the zenodo metadata should now properly match the JOSS paper metadata. Let me know if anything is missing- thanks!
